Moving base classes of file backends to base/
