+++ ibis/backends/hdf5/tests/test_hdf5.py
@@ -3,7 +3,7 @@ import pytest
 from pandas.util import testing as tm
 
 import ibis
-from ibis.backends.base_file import FileDatabase
+from ibis.backends.base.file import FileDatabase
 from ibis.backends.hdf5 import HDFTable
