+++ ibis/backends/parquet/__init__.py
@@ -10,7 +10,7 @@ import ibis.expr.operations as ops
 import ibis.expr.schema as sch
 import ibis.expr.types as ir
 from ibis.backends.base import BaseBackend
-from ibis.backends.base_file import FileClient
+from ibis.backends.base.file import FileClient
 from ibis.backends.pandas import PandasExprTranslator
 from ibis.backends.pandas.core import execute, execute_node
