+++ ibis/backends/parquet/tests/test_parquet.py
@@ -6,7 +6,7 @@ import pytest
 from pandas.util import testing as tm
 
 import ibis
-from ibis.backends.base_file import FileDatabase
+from ibis.backends.base.file import FileDatabase
 from ibis.backends.parquet import ParquetTable
 
 pytestmark = pytest.mark.skipif(
